Move macros into module
