+++ server/src/dbnet/connection.rs
@@ -91,79 +91,6 @@ pub mod prelude {
     pub use crate::registry;
     pub use crate::util::Unwrappable;
     pub use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    #[macro_export]
-    macro_rules! kve {
-        ($con:expr, $store:expr) => {
-            match $store.get_kvstore() {
-                Ok(store) => store,
-                _ => {
-                    // wrong model
-                    return $con
-                        .write_response(crate::protocol::responses::groups::WRONG_MODEL)
-                        .await;
-                }
-            }
-        };
-    }
-    #[macro_export]
-    macro_rules! default_keyspace {
-        ($store:expr, $con:expr) => {
-            match $store.get_keyspace() {
-                Ok(ks) => ks,
-                Err(_) => {
-                    return $con
-                        .write_response(crate::protocol::responses::groups::DEFAULT_UNSET)
-                        .await;
-                }
-            }
-        };
-    }
-    #[macro_export]
-    macro_rules! conwrite {
-        ($con:expr, $what:expr) => {
-            $con.write_response($what).await
-        };
-    }
-    #[macro_export]
-    macro_rules! aerr {
-        ($con:expr, aerr) => {
-            return conwrite!($con, crate::protocol::responses::groups::ACTION_ERR)
-        };
-    }
-    #[macro_export]
-    macro_rules! get_tbl {
-        ($entity:expr, $store:expr, $con:expr) => {{
-            use crate::corestore::memstore::DdlError;
-            match $store.get_table($entity) {
-                Ok(tbl) => tbl,
-                Err(DdlError::DefaultNotFound) => {
-                    return conwrite!($con, crate::protocol::responses::groups::DEFAULT_UNSET);
-                }
-                Err(DdlError::ObjectNotFound) => {
-                    return conwrite!(
-                        $con,
-                        crate::protocol::responses::groups::CONTAINER_NOT_FOUND
-                    );
-                }
-                Err(_) => unsafe { impossible!() },
-            }
-        }};
-        ($store:expr, $con:expr) => {{
-            match $store.get_ctable() {
-                Some(tbl) => tbl,
-                None => return conwrite!($con, crate::protocol::responses::groups::DEFAULT_UNSET),
-            }
-        }};
-    }
-    #[macro_export]
-    macro_rules! handle_entity {
-        ($con:expr, $ident:expr) => {{
-            match crate::queryengine::parser::get_query_entity(&$ident) {
-                Ok(e) => e,
-                Err(e) => return conwrite!($con, e),
-            }
-        }};
-    }
 }
 
 /// # The `ProtocolConnectionExt` trait
